Add truncate method to DeleteFiles and support fast truncate table
